feat: Refactor task event data structures and update Python version.
