+++ src/avro/events/generate_tasks_event.py
@@ -1,47 +1,21 @@
 from dataclasses import dataclass
 from typing import List, Optional
 
-from src.avro.enums.rarity import Rarity
-from src.avro.enums.task_topic import TaskTopic
 
-
-@dataclass
-class GenerateTask:
-    taskId: str
-    version: int = 0
-    rarity: Optional[Rarity] = None
-    topics: Optional[List[TaskTopic]] = None
-
-    def to_dict(self) -> dict:
-        result = {
-            "taskId": self.taskId,
-            "version": self.version,
-            "rarity": self.rarity.value if self.rarity else None,
-            "topics": [t.value for t in self.topics] if self.topics is not None else [],
-        }
-        return result
-
-    @classmethod
-    def from_dict(cls, data: dict) -> "GenerateTask":
-        return cls(
-            taskId=data["taskId"],
-            version=data.get("version", 0),
-            rarity=Rarity(data["rarity"]) if data.get("rarity") else None,
-            topics=[TaskTopic(t) for t in data.get("topics", [])],
-        )
+from src.avro.events.task import Task
 
 
 @dataclass
 class GenerateTasksEvent:
     txId: Optional[str] = None
-    playerId: Optional[int] = None
-    inputs: Optional[List[GenerateTask]] = None
+    userId: Optional[int] = None
+    tasks: Optional[List[Task]] = None
 
     def to_dict(self) -> dict:
         result = {
             "txId": self.txId,
-            "playerId": self.playerId,
-            "inputs": [task.to_dict() for task in self.inputs] if self.inputs else [],
+            "userId": self.userId,
+            "tasks": [task.to_dict() for task in self.tasks] if self.tasks else [],
         }
         return result
 
@@ -49,10 +23,8 @@ class GenerateTasksEvent:
     def from_dict(cls, data: dict) -> "GenerateTasksEvent":
         return cls(
             txId=data.get("txId"),
-            playerId=data.get("playerId"),
-            inputs=(
-                [GenerateTask.from_dict(t) for t in data["inputs"]]
-                if data.get("inputs")
-                else []
+            userId=data.get("userId"),
+            tasks=(
+                [Task.from_dict(t) for t in data["tasks"]] if data.get("tasks") else []
             ),
         )
