+++ src/kafka/consumer.py
@@ -13,8 +13,9 @@ from src.kafka.interceptors import ConsumerLocaleInterceptor
 from src.kafka.producer import send_save_tasks_event
 from src.services.task_service import TaskService
 from src.services.avro_serialization import ConfluentAvroService
-from src.avro.events.save_tasks_event import SaveTask, SaveTasksEvent
-from src.avro.events.generate_tasks_event import GenerateTask, GenerateTasksEvent
+from src.avro.events.save_tasks_event import SaveTasksEvent
+from src.avro.events.generate_tasks_event import GenerateTasksEvent
+from src.avro.events.task import Task
 from src.avro.enums.rarity import Rarity
 from src.avro.enums.task_topic import TaskTopic
 from src.config.config_loader import (
@@ -60,29 +61,29 @@ def register_consumers(broker: KafkaBroker):
 
             event = GenerateTasksEvent.from_dict(event_dict)
 
-            if not event.inputs:
+            if not event.tasks:
                 logger.warning(
-                    f"Received empty task list for player {event.playerId}, txId={event.txId}"
+                    f"Received empty task list for player {event.userId}, txId={event.txId}"
                 )
                 return
 
             logger.info(
-                f"Received GenerateTasksEvent: playerId={event.playerId}, "
-                f"txId={event.txId}, tasks_count={len(event.inputs)}"
+                f"Received GenerateTasksEvent: userId={event.userId}, "
+                f"txId={event.txId}, tasks_count={len(event.tasks)}"
             )
 
             if is_feature_enabled("debug_logging"):
                 logger.debug(f"Full event data: {event}")
 
             # Группируем задачи по (topics, rarity) для batch-обработки
-            task_groups = group_tasks_by_params(event.inputs)
+            task_groups = group_tasks_by_params(event.tasks)
 
             logger.info(
-                f"Grouped {len(event.inputs)} tasks into {len(task_groups)} batch groups"
+                f"Grouped {len(event.tasks)} tasks into {len(task_groups)} batch groups"
             )
 
             # Обрабатываем каждую группу батчем
-            save_tasks: List[SaveTask] = []
+            save_tasks: List[Task] = []
 
             # В handle_task_request(), после группировки:
 
@@ -107,11 +108,11 @@ def register_consumers(broker: KafkaBroker):
                     )
 
                     task_input = group_tasks[0]
-                    save_task = SaveTask.from_generated(task_input, generated_task)
+                    save_task = Task.from_generated(task_input, generated_task)
                     save_tasks.append(save_task)
 
                     logger.info(
-                        f"Generated single task '{generated_task.title.en}' -> taskId={task_input.taskId}"
+                        f"Generated single task '{generated_task.title.en}' -> taskId={task_input.id}"
                     )
                 else:
                     # Batch-генерация для 2+ задач
@@ -126,11 +127,11 @@ def register_consumers(broker: KafkaBroker):
 
                     # Мапим сгенерированные задачи на исходные taskId
                     for task_input, generated_task in zip(group_tasks, generated_tasks):
-                        save_task = SaveTask.from_generated(task_input, generated_task)
+                        save_task = Task.from_generated(task_input, generated_task)
                         save_tasks.append(save_task)
 
                         logger.info(
-                            f"Mapped generated task '{generated_task.title.en}' -> taskId={task_input.taskId}"
+                            f"Mapped generated task '{generated_task.title.en}' -> taskId={task_input.id}"
                         )
 
             logger.info(
@@ -139,13 +140,13 @@ def register_consumers(broker: KafkaBroker):
 
             # Отправляем результат
             save_event = SaveTasksEvent(
-                txId=event.txId, playerId=event.playerId, tasks=save_tasks
+                txId=event.txId, userId=event.userId, tasks=save_tasks
             )
 
             await send_save_tasks_event(producer, save_event)
 
             logger.info(
-                f"Published SaveTasksEvent: playerId={event.playerId}, "
+                f"Published SaveTasksEvent: userId={event.userId}, "
                 f"txId={event.txId}, tasks_count={len(save_tasks)}"
             )
 
@@ -156,22 +157,20 @@ def register_consumers(broker: KafkaBroker):
 
 
 def group_tasks_by_params(
-    inputs: List[GenerateTask],
-) -> Dict[Tuple[Tuple[TaskTopic, ...], Rarity], List[GenerateTask]]:
+    tasks: List[Task],
+) -> Dict[Tuple[Tuple[TaskTopic, ...], Rarity], List[Task]]:
     """
     Группирует задачи по (topics, rarity) для batch-обработки.
 
     Args:
-        inputs: Список входящих задач
+        tasks: Список входящих задач
 
     Returns:
         Словарь {(topics_tuple, rarity): [список задач с этими параметрами]}
     """
-    groups: Dict[Tuple[Tuple[TaskTopic, ...], Rarity], List[GenerateTask]] = (
-        defaultdict(list)
-    )
+    groups: Dict[Tuple[Tuple[TaskTopic, ...], Rarity], List[Task]] = defaultdict(list)
 
-    for task_input in inputs:
+    for task_input in tasks:
         # Создаем ключ для группировки
         topics_tuple = tuple(sorted(task_input.topics or [], key=lambda t: t.value))
         rarity = task_input.rarity if task_input.rarity is not None else Rarity.COMMON
